requirements-production.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-axes==8.0.0
django-cors-headers==4.2.0
django-defender==0.9.8
django-filter==25.2
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==21.2.0
inflection==0.5.1
kombu==5.5.4
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.5.0
